# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Catálogo Publiex Maracaibo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads/complete:
    post:
      operationId: completeSignedUpload
      tags: [products]
      summary: Verify a completed Storage upload
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadCompleteInput"
      responses:
        "200":
          description: Verified image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImagesResponseImagesItem"
        "400":
          description: Invalid or missing image
        "401":
          description: Authentication required
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active catalog banners
      responses:
        "200":
          description: Ordered catalog banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: uploadBanners
      tags: [banners]
      summary: Upload catalog banners
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Banners uploaded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
        "400":
          description: Invalid images
        "401":
          description: Authentication required
        "413":
          description: Upload too large
  /banners/from-storage:
    post:
      operationId: createBannersFromStorage
      tags: [banners]
      summary: Register verified banner uploads
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageBannersInput"
      responses:
        "201":
          description: Registered banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
        "400":
          description: Invalid uploads
        "401":
          description: Authentication required
  /banners/order:
    patch:
      operationId: reorderBanners
      tags: [banners]
      summary: Reorder catalog banners
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerOrder"
      responses:
        "200":
          description: Updated banner order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
        "400":
          description: Invalid order
        "401":
          description: Authentication required
  /banners/{id}:
    patch:
      operationId: updateBannerLink
      tags: [banners]
      summary: Link a banner to a promotional product
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBannerBody"
      responses:
        "200":
          description: Banner with updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
        "400":
          description: Invalid or non-promotional product
        "401":
          description: Authentication required
        "404":
          description: Banner not found
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a catalog banner
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Banner deleted
        "400":
          description: Invalid banner id
        "401":
          description: Authentication required
        "404":
          description: Banner not found
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of reusable categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a product category
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category
        "401":
          description: Authentication required
        "409":
          description: Category already exists
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a product category and optionally reassign its products
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCategoryBody"
      responses:
        "200":
          description: Category deleted and products reassigned when necessary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCategoryResponse"
        "400":
          description: Invalid replacement category
        "401":
          description: Authentication required
        "404":
          description: Category not found
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 1
          maxLength: 80
        slug:
          type: string
          minLength: 1
          maxLength: 160
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
    DeleteCategoryBody:
      type: object
      required: [replacementCategoryId]
      properties:
        replacementCategoryId:
          type: integer
          minimum: 1
          nullable: true
          description: Existing category that will receive products from the deleted category
    DeleteCategoryResponse:
      type: object
      required: [deletedCategoryId, replacementCategoryId, movedProductCount]
      properties:
        deletedCategoryId:
          type: integer
        replacementCategoryId:
          type: integer
          nullable: true
        movedProductCount:
          type: integer
          minimum: 0
    Banner:
      type: object
      required: [id, url, position, createdAt, productId, productName, link]
      properties:
        id:
          type: integer
        url:
          type: string
        position:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        productId:
          type: integer
          nullable: true
        productName:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
    UpdateBannerBody:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
          minimum: 1
          nullable: true
    BannerOrder:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: integer
            minimum: 1
...
    UploadImagesResponseImagesItem:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
    UploadSignInput:
      type: object
      required: [kind, name, contentType, size]
      properties:
        kind:
          type: string
          enum: [product, banner]
        name:
          type: string
          maxLength: 255
        contentType:
          type: string
          enum: [image/avif, image/gif, image/jpeg, image/png, image/webp]
        size:
          type: integer
          minimum: 1
          maximum: 5242880
...
    UploadCompleteInput:
      allOf:
        - $ref: "#/components/schemas/UploadSignInput"
        - type: object
          required: [storagePath]
          properties:
            storagePath:
              type: string
              maxLength: 300
    StorageBannerUpload:
      type: object
      required: [storagePath, name, contentType, size]
      properties:
        storagePath:
          type: string
          maxLength: 300
        name:
          type: string
          maxLength: 255
        contentType:
          type: string
          enum: [image/avif, image/gif, image/jpeg, image/png, image/webp]
        size:
          type: integer
          minimum: 1
          maximum: 5242880
    StorageBannersInput:
      type: object
      required: [uploads]
      properties:
        uploads:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: "#/components/schemas/StorageBannerUpload"
...